requirements.txt
glob3
pyserial
requests